+++ pkg/compiler/testdata/bar/bar.go
