+++ pkg/compiler/testdata/foo/foo.go
