+++ pkg/compiler/testdata/foobar/bar.go
