+++ pkg/compiler/testdata/foobar/foo.go
